move scripts and add readme c# extensions/utils
