+++ C#/Extensions/AssemblyExtensions.cs
