+++ C#/Extensions/ByteArrayExtensions.cs
