+++ C#/Extensions/ByteExtensions.cs
