+++ C#/Extensions/StringExtensions.cs
