+++ C#/Utils/HashUtils.cs
