+++ C#/Utils/MiscUtils.cs
