--- UNITY/Utils/readme.md
@@ -1,5 +1,0 @@
-A lot of useful script for Unity or just C# you can compile it to make a dll lib et juste use .CS file directly in your project<br><br>
-
-
-# HOW TO USE:
-For Unity .Net 2.xx && 4.xx && IL2CPP or Visual studio all version
